Refactor/cleanup performance tests
Prior code was very outdated and not at all useful.

With these changes, I was able to get a "load test" running against
my local devstack setup, with 20 concurrent users.
