+++ requirements/perf.txt
@@ -1,3 +1,3 @@
-locustio==0.7.0
+locustio==0.7.3
 loremipsum==1.0.5
 pyzmq==14.0.1
