+++ test/performance/base.py
@@ -1,5 +1,5 @@
 """
-Performance tests for the OpenAssessment XBlock.
+Performance test utility for the OpenAssessment XBlock.
 """
 
 import os
@@ -8,8 +8,6 @@ import random
 from collections import namedtuple
 import gevent
 import loremipsum
-from locust import HttpLocust, TaskSet, task
-
 
 
 class OpenAssessmentPage(object):
@@ -25,9 +23,9 @@ class OpenAssessmentPage(object):
 
     PROBLEMS = {
         'peer_then_self': ProblemFixture(
-            course_id="ora2/1/1",
-            base_url= "courses/ora2/1/1/courseware/efa85eb090164a208d772a344df7181d/69f15a02c5af4e95b9c5525771b8f4ee/",
-            base_handler_url="courses/ora2/1/1/xblock/i4x:;_;_ora2;_1;_openassessment;_0e2bbf6cc89e45d98b028fa4e2d46314/handler/",
+            course_id="course-v1:edx+ORA203+course",
+            base_url="courses/course-v1:edx+ORA203+course/xblock/block-v1:edx+ORA203+course+type@openassessment+block@47dc34e528f441f493db14a2cbdfa8b9/",
+            base_handler_url="courses/course-v1:edx+ORA203+course/xblock/block-v1:edx+ORA203+course+type@openassessment+block@47dc34e528f441f493db14a2cbdfa8b9/handler/",
             rubric_options={
                 'Ideas': ['Poor', 'Fair', 'Good'],
                 'Content': ['Poor', 'Fair', 'Good', 'Excellent']
@@ -36,17 +34,6 @@ class OpenAssessmentPage(object):
                 'render_submission', 'render_peer_assessment',
                 'render_self_assessment', 'render_grade',
             ]
-        ),
-
-        'example_based': ProblemFixture(
-            course_id="ora2/1/1",
-            base_url="courses/ora2/1/1/courseware/efa85eb090164a208d772a344df7181d/fb039ef8a34641509190918ada79122a/",
-            base_handler_url="courses/ora2/1/1/xblock/i4x:;_;_ora2;_1;_openassessment;_8df3fa4de26747e0ad99b4157e45f5e5/handler/",
-            rubric_options={
-                'Ideas': ['Bad', 'Good'],
-                'Content': ['Bad', 'Good']
-            },
-            render_step_handlers=['render_submission', 'render_grade']
         )
     }
 
@@ -105,7 +92,7 @@ class OpenAssessmentPage(object):
         Submit a response.
         """
         payload = json.dumps({
-            'submission': u' '.join(loremipsum.get_paragraphs(random.randint(1, 10))),
+            'submission': [u' '.join(loremipsum.get_sentence())],
         })
         self.client.post(self.handler_url('submit'), data=payload, headers=self._post_headers, verify=False)
 
@@ -171,78 +158,3 @@ class OpenAssessmentPage(object):
             'X-CSRFToken': self.client.cookies.get('csrftoken', ''),
             'Referer': self.hostname
         }
-
-
-class OpenAssessmentTasks(TaskSet):
-    """
-    Virtual user interactions with the OpenAssessment XBlock.
-    """
-
-    def __init__(self, *args, **kwargs):  # pylint: disable=W0613
-        """
-        Initialize the task set.
-        """
-        super(OpenAssessmentTasks, self).__init__(*args, **kwargs)
-        self.hostname = self.locust.host
-        self.page = None
-
-    @task
-    def peer_and_self(self):
-        """
-        Test the peer-->self workflow.
-        """
-        if self.page is None:
-            self.page = OpenAssessmentPage(self.hostname, self.client, 'peer_then_self')  # pylint: disable=E1101
-            self.page.log_in()
-
-        if not self.page.logged_in:
-            self.page.log_in()
-        else:
-            self._submit_response()
-
-            # Randomly peer/self assess or log in as a new user.
-            # This should be sufficient to get students through
-            # the entire flow (satisfying the requirements for peer assessment).
-            action = random.randint(0, 100)
-            if action <= 80:
-                continue_grading = random.randint(0, 10) < 4
-                self.page.peer_assess(continue_grading=continue_grading)
-                self.page.self_assess()
-            else:
-                self.page.log_in()
-
-    @task
-    def example_based(self):
-        """
-        Test example-based assessment only.
-        """
-        if self.page is None:
-            self.page = OpenAssessmentPage(self.hostname, self.client, 'example_based')  # pylint: disable=E1101
-            self.page.log_in()
-
-        if not self.page.logged_in:
-            self.page.log_in()
-        else:
-            self._submit_response()
-            if random.randint(0, 100) < 50:
-                self.page.log_in()
-
-    def _submit_response(self):
-        """
-        Simulate the user loading the page, submitting a response,
-        then reloading the steps (usually triggered by AJAX).
-        If the user has already submitted, the handler will return
-        an error message in the JSON, but the HTTP status will still be 200.
-        """
-        self.page.load_steps()
-        self.page.submit_response()
-        self.page.load_steps()
-
-
-class OpenAssessmentLocust(HttpLocust):
-    """
-    Performance test definition for the OpenAssessment XBlock.
-    """
-    task_set = OpenAssessmentTasks
-    min_wait = 10000
-    max_wait = 15000
